Crate google_firestore1_beta1

This documentation was generated from Firestore crate version 1.0.9+20190419, where 20190419 is the exact revision of the firestore:v1beta1 schema built by the mako code generator v1.0.9.

Everything else about the Firestore v1_beta1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

projects

databases documents batch get, databases documents begin transaction, databases documents commit, databases documents create document, databases documents delete, databases documents get, databases documents list, databases documents list collection ids, databases documents listen, databases documents patch, databases documents rollback, databases documents run query, databases documents write, databases export documents, databases import documents, databases indexes create, databases indexes delete, databases indexes get and databases indexes list

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.projects().databases_import_documents(...).doit()
let r = hub.projects().databases_export_documents(...).doit()
let r = hub.projects().databases_indexes_create(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-firestore1_beta1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_firestore1_beta1 as firestore1_beta1;
use firestore1_beta1::GoogleFirestoreAdminV1beta1ImportDocumentsRequest;
use firestore1_beta1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use firestore1_beta1::Firestore;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Firestore::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = GoogleFirestoreAdminV1beta1ImportDocumentsRequest::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().databases_import_documents(req, "name")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

ArrayValue
An array value.

BatchGetDocumentsRequest
The request for Firestore.BatchGetDocuments.

BatchGetDocumentsResponse
The streamed response for Firestore.BatchGetDocuments.

BeginTransactionRequest
The request for Firestore.BeginTransaction.

BeginTransactionResponse
The response for Firestore.BeginTransaction.

Chunk

CollectionSelector
A selection of a collection, such as messages as m1.

CommitRequest
The request for Firestore.Commit.

CommitResponse
The response for Firestore.Commit.

CompositeFilter
A filter that merges multiple other filters using the given operator.

ContentRange
Implements the Content-Range header, for serialization only

Cursor
A position in a query result set.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Document
A Firestore document.

DocumentChange
A Document has changed.

DocumentDelete
A Document has been deleted.

DocumentMask
A set of field paths on a document. Used to restrict a get or update operation on a document to a subset of its fields. This is different from standard field masks, as this is always scoped to a Document, and takes in account the dynamic nature of Value.

DocumentRemove
A Document has been removed from the view of the targets.

DocumentTransform
A transformation of a document.

DocumentsTarget
A target specified by a set of documents names.

DummyNetworkStream

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

ExistenceFilter
A digest of all the documents that match a given target.

FieldFilter
A filter on a specific field.

FieldReference
A reference to a field, such as max(messages.time) as max_time.

FieldTransform
A transformation of a field of the document.

Filter
A filter.

Firestore
Central instance to access all Firestore related resource activities

GoogleFirestoreAdminV1beta1IndexField
A field of an index.

GoogleFirestoreAdminV1beta1ImportDocumentsRequest
The request for FirestoreAdmin.ImportDocuments.

GoogleFirestoreAdminV1beta1ExportDocumentsRequest
The request for FirestoreAdmin.ExportDocuments.

GoogleFirestoreAdminV1beta1ListIndexesResponse
The response for FirestoreAdmin.ListIndexes.

GoogleFirestoreAdminV1beta1Index
An index definition.

GoogleLongrunningOperation
This resource represents a long-running operation that is the result of a network API call.

JsonServerError
A utility type which can decode a server response that indicates error

LatLng
An object representing a latitude/longitude pair. This is expressed as a pair of doubles representing degrees latitude and degrees longitude. Unless specified otherwise, this must conform to the WGS84 standard. Values must be within normalized ranges.

ListCollectionIdsRequest
The request for Firestore.ListCollectionIds.

ListCollectionIdsResponse
The response from Firestore.ListCollectionIds.

ListDocumentsResponse
The response for Firestore.ListDocuments.

ListenRequest
A request for Firestore.Listen

ListenResponse
The response for Firestore.Listen.

MapValue
A map value.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Order
An order on a field.

Precondition
A precondition on a document, used for conditional operations.

ProjectDatabaseDocumentBatchGetCall
Gets multiple documents.

ProjectDatabaseDocumentBeginTransactionCall
Starts a new transaction.

ProjectDatabaseDocumentCommitCall
Commits a transaction, while optionally updating documents.

ProjectDatabaseDocumentCreateDocumentCall
Creates a new document.

ProjectDatabaseDocumentDeleteCall
Deletes a document.

ProjectDatabaseDocumentGetCall
Gets a single document.

ProjectDatabaseDocumentListCall
Lists documents.

ProjectDatabaseDocumentListCollectionIdCall
Lists all the collection IDs underneath a document.

ProjectDatabaseDocumentListenCall
Listens to changes.

ProjectDatabaseDocumentPatchCall
Updates or inserts a document.

ProjectDatabaseDocumentRollbackCall
Rolls back a transaction.

ProjectDatabaseDocumentRunQueryCall
Runs a query.

ProjectDatabaseDocumentWriteCall
Streams batches of document updates and deletes, in order.

ProjectDatabaseExportDocumentCall
Exports a copy of all or a subset of documents from Google Cloud Firestore to another storage system, such as Google Cloud Storage. Recent updates to documents may not be reflected in the export. The export occurs in the background and its progress can be monitored and managed via the Operation resource that is created. The output of an export may only be used once the associated operation is done. If an export operation is cancelled before completion it may leave partial data behind in Google Cloud Storage.

ProjectDatabaseImportDocumentCall
Imports documents into Google Cloud Firestore. Existing documents with the same name are overwritten. The import occurs in the background and its progress can be monitored and managed via the Operation resource that is created. If an ImportDocuments operation is cancelled, it is possible that a subset of the data has already been imported to Cloud Firestore.

ProjectDatabaseIndexeCreateCall
Creates the specified index. A newly created index's initial state is CREATING. On completion of the returned google.longrunning.Operation, the state will be READY. If the index already exists, the call will return an ALREADY_EXISTS status.

ProjectDatabaseIndexeDeleteCall
Deletes an index.

ProjectDatabaseIndexeGetCall
Gets an index.

ProjectDatabaseIndexeListCall
Lists the indexes that match the specified filters.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Firestore hub.

Projection
The projection of document's fields to return.

QueryTarget
A target specified by a query.

RangeResponseHeader

ReadOnly
Options for a transaction that can only be used to read documents.

ReadWrite
Options for a transaction that can be used to read and write documents.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

RollbackRequest
The request for Firestore.Rollback.

RunQueryRequest
The request for Firestore.RunQuery.

RunQueryResponse
The response for Firestore.RunQuery.

ServerError

ServerMessage

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. The error model is designed to be:

StructuredQuery
A Firestore query.

Target
A specification of a set of documents to listen to.

TargetChange
Targets being watched have changed.

TransactionOptions
Options for creating a new transaction.

UnaryFilter
A filter with a single operand.

Value
A message that can hold any of the supported value types.

Write
A write on a document.

WriteRequest
The request for Firestore.Write.

WriteResponse
The response for Firestore.Write.

WriteResult
The result of applying a write.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.